OneDrive/Documents/Tech-news-agent/requirements.txt
openai>=0.27.0
requests>=2.28.0
python-dotenv>=0.19.0
schedule>=1.1.0
